# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/instagram/oauth/authorize": {
      "get": {
        "operationId": "InstagramController_initiateOAuth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns authorization URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorizationUrl": {
                      "type": "string",
                      "example": "https://api.instagram.com/oauth/authorize?..."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate Instagram OAuth flow",
        "tags": [
          "Instagram Integration"
        ]
      }
    },
    "/api/instagram/oauth/callback": {
      "get": {
        "operationId": "InstagramController_handleOAuthCallback",
        "parameters": [
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Authorization code from Instagram",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "CSRF state token",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": false,
            "in": "query",
            "description": "Error code if user denied access",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to frontend"
          }
        },
        "summary": "Handle Instagram OAuth callback",
        "tags": [
          "Instagram Integration"
        ]
      }
    },
    "/api/instagram/accounts/{id}/sync": {
      "post": {
        "operationId": "InstagramAccountController_syncAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sync account metadata from Instagram",
        "tags": [
          "Instagram Account Management"
        ]
      }
    },
    "/api/instagram/accounts/{id}/disconnect": {
      "delete": {
        "operationId": "InstagramController_disconnectAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account disconnected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Instagram account disconnected successfully"
                    },
                    "accountId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disconnect Instagram account",
        "tags": [
          "Instagram Integration"
        ]
      }
    },
    "/api/instagram/system/accounts": {
      "get": {
        "description": "Lists all Instagram Business Accounts accessible via the system user token",
        "operationId": "InstagramController_listAvailableAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of available Instagram Business Accounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableAccountsResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "System user token not configured or API error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List available Instagram Business Accounts",
        "tags": [
          "Instagram Integration"
        ]
      }
    },
    "/api/instagram/system/accounts/{id}": {
      "get": {
        "description": "Retrieves detailed information about a specific Instagram Business Account",
        "operationId": "InstagramController_getAccountDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Instagram Business Account ID",
            "schema": {
              "example": "17841405309211844",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instagram Business Account details"
          },
          "404": {
            "description": "Account not found or not accessible"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Instagram Business Account details",
        "tags": [
          "Instagram Integration"
        ]
      }
    },
    "/api/instagram/system/accounts/link": {
      "post": {
        "description": "Links an Instagram Business Account to the authenticated user as a client account",
        "operationId": "InstagramController_linkInstagramAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkInstagramAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account linked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkAccountResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Instagram Business Account ID or already linked"
          },
          "404": {
            "description": "Instagram Business Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Link Instagram Business Account to client",
        "tags": [
          "Instagram Integration"
        ]
      }
    },
...
  },
  "info": {
    "title": "Social Selling API",
    "description": "\n# Social Selling Platform API\n\nComplete REST API for Instagram-based social selling platform.\n\n## Features\n- Instagram OAuth integration\n- Product catalog management\n- Direct message automation\n- Post scheduling and analytics\n- Multi-account support\n\n## Authentication\nAll authenticated endpoints require a Bearer token in the Authorization header.\nObtain tokens via POST /api/auth/login endpoint.\n\n## Rate Limiting\n- Authenticated: 1000 requests/hour\n- Unauthenticated: 100 requests/hour\n\n## Session Management\nSessions are managed via HTTP-only cookies and can be tracked per device.\n  ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.socialselling.willianbvsanches.com",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.socialselling.willianbvsanches.com",
      "description": "Production Environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT access token",
        "in": "header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ssell_session",
        "description": "Session cookie for authenticated requests"
      }
    },
    "schemas": {
...
      "InstagramBusinessAccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Instagram Business Account ID",
            "example": "17841405309211844"
          },
          "username": {
            "type": "string",
            "description": "Instagram username",
            "example": "my_business_account"
          },
          "name": {
            "type": "string",
            "description": "Account display name",
            "example": "My Business"
          },
          "profile_picture_url": {
            "type": "string",
            "description": "Profile picture URL"
          },
          "followers_count": {
            "type": "number",
            "description": "Number of followers",
            "example": 1500
          },
          "follows_count": {
            "type": "number",
            "description": "Number of accounts following",
            "example": 300
          },
          "media_count": {
            "type": "number",
            "description": "Number of media posts",
            "example": 125
          },
          "biography": {
            "type": "string",
            "description": "Account biography"
          },
          "website": {
            "type": "string",
            "description": "Website URL"
          },
          "facebookPageId": {
            "type": "string",
            "description": "Connected Facebook Page ID",
            "example": "123456789012345"
          },
          "facebookPageName": {
            "type": "string",
            "description": "Connected Facebook Page Name",
            "example": "My Business Page"
          }
        },
        "required": [
          "id",
          "username",
          "facebookPageId",
          "facebookPageName"
        ]
      },
      "AvailableAccountsResponseDto": {
        "type": "object",
        "properties": {
          "accounts": {
            "description": "List of available Instagram Business Accounts",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstagramBusinessAccountDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of accounts",
            "example": 3
          }
        },
        "required": [
          "accounts",
          "total"
        ]
      },
      "LinkInstagramAccountDto": {
        "type": "object",
        "properties": {
          "instagramBusinessAccountId": {
            "type": "string",
            "description": "Instagram Business Account ID to link",
            "example": "17841405309211844"
          },
          "facebookPageId": {
            "type": "string",
            "description": "Optional: Facebook Page ID (for verification)",
            "example": "123456789012345"
          }
        },
        "required": [
          "instagramBusinessAccountId"
        ]
      },
      "LinkAccountResponseDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "Client account ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "username": {
            "type": "string",
            "description": "Instagram username",
            "example": "my_business_account"
          },
          "instagramBusinessAccountId": {
            "type": "string",
            "description": "Instagram Business Account ID",
            "example": "17841405309211844"
          },
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Instagram Business Account linked successfully"
          }
        },
        "required": [
          "accountId",
          "username",
          "instagramBusinessAccountId",
          "message"
        ]
      },
      "AccountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Account UUID"
          },
          "platform": {
            "type": "string",
            "description": "Platform name",
            "example": "instagram"
          },
          "username": {
            "type": "string",
            "description": "Instagram username"
          },
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "profilePictureUrl": {
            "type": "string",
            "description": "Profile picture URL"
          },
          "followerCount": {
            "type": "number",
            "description": "Follower count"
          },
          "followingCount": {
            "type": "number",
            "description": "Following count"
          },
          "mediaCount": {
            "type": "number",
            "description": "Media count"
          },
          "biography": {
            "type": "string",
            "description": "Biography"
          },
          "website": {
            "type": "string",
            "description": "Website URL"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "token_expired",
              "disconnected",
              "rate_limited",
              "error"
            ],
            "description": "Account status"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "personal",
              "business",
              "creator"
            ],
            "description": "Account type"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "lastSyncAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last sync timestamp"
          },
          "tokenExpiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Token expiration timestamp"
          }
        },
        "required": [
          "id",
          "platform",
          "username",
          "status",
          "accountType",
          "createdAt"
        ]
      },
...
    }
  }
}